tests/requirements.txt
# Integration testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-html==4.1.1
requests==2.31.0
psycopg2-binary==2.9.9
redis==5.0.1
selenium==4.15.2
webdriver-manager==4.0.1
aiohttp==3.9.1
pydantic==2.5.0
python-dotenv==1.0.0
faker==20.1.0
pytest-xdist==3.5.0
pytest-cov==4.1.0
pytest-timeout==2.2.0
